fixed malderbot example
