+++ examples/app/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(notes)
 add_subdirectory(minesweeper)
+add_subdirectory(fractal)
