+++ examples/app/fractal/README.md
@@ -1,0 +1,9 @@
+# Fractal Example
+
+@auiexample{app}
+Fractal viewer application demonstrating usage of custom shaders.
+
+![Fractal](https://sun9-42.userapi.com/impf/WruyOdMmMBrRfpjJ7QrhFepZj7obL3VMGxNSaw/Tr8XxKqdVV8.jpg?size=1261x740&quality=96&proxy=1&sign=f6b851a26a7c40a5f1c22367a34f4c71&type=album)
+
+If you are familiar with OpenGL, you can use custom shaders with AUI. In this example, we're using such shader to
+display a GPU-computed Mandelbrot set.
