+++ examples/app/fractal/assets/img/color_scheme_wikipedia.png
