+++ examples/app/fractal/src/FractalWindow.h
@@ -11,12 +11,9 @@
 
 #pragma once
 
-
 #include <AUI/Platform/AWindow.h>
-#include "FractalView.h"
 
-class JumpToCoordsWindow: public AWindow {
+class FractalWindow : public AWindow {
 public:
-    explicit JumpToCoordsWindow(_<FractalView> fractalView, AWindow* parent);
+    FractalWindow();
 };
-
