+++ examples/app/fractal/src/main.cpp
@@ -12,8 +12,7 @@
 #include <AUI/Platform/Entry.h>
 #include "FractalWindow.h"
 
-AUI_ENTRY
-{
+AUI_ENTRY {
     _new<FractalWindow>()->show();
-	return 0;
+    return 0;
 }
